Added logo and fonts, included react-router in the project and created some routes, added styling.
