+++ src/components/category-item/caterory-item.component.jsx
@@ -2,7 +2,7 @@ import './category-item.styles.scss';
 
 const CategoryItem = ({ category }) => {
   const { title, imageUrl } = category;
-  console.log(category);
+
   return (
     <div className="category-container">
       <div
