+++ src/routes/home/home.component.jsx
@@ -1,0 +1,42 @@
+import { Outlet } from 'react-router-dom';
+
+import CategoriesContainer from '../../components/categories-container/categories-container.component';
+
+const Home = () => {
+  const categories = [
+    {
+      id: 1,
+      title: 'hats',
+      imageUrl: 'https://i.ibb.co/cvpntL1/hats.png',
+    },
+    {
+      id: 2,
+      title: 'jackets',
+      imageUrl: 'https://i.ibb.co/px2tCc3/jackets.png',
+    },
+    {
+      id: 3,
+      title: 'footware',
+      imageUrl: 'https://i.ibb.co/0jqHpnp/sneakers.png',
+    },
+    {
+      id: 4,
+      title: 'womens',
+      imageUrl: 'https://i.ibb.co/GCCdy8t/womens.png',
+    },
+    {
+      id: 5,
+      title: 'mens',
+      imageUrl: 'https://i.ibb.co/R70vBrQ/men.png',
+    },
+  ];
+
+  return (
+    <div>
+      <CategoriesContainer categories={categories} />
+      <Outlet />
+    </div>
+  );
+};
+
+export default Home;
